refactor `lds_lib` to use `jax.lax.scan` (#590)
* refactor: linreg_eb_modelsel_vs_n.py

* refactor: kf_parallel_demo.py, kf_tracking_demo.py

Force float type for the parameters

* refactor: lds_lib.py

* Replace custom Kalman filter loop in favour of `jax.lax.scan`

* feat: lds_lib.py

Define `__smother_step` (to be used with jax.lax.scan)

* refactor: lds_lib.py

Modify `__kalman_smoother` method training loop for jax.lax.scan

* fix: kf_spiral_demo.py

Force float type for initial state

* create: lds_cts_time_lib

Move ContinuousKalmanFilter class from lds_lib to its custom class

* refactor: kf_continuous_circle_demo

Make use of `lds_cts_time_lib`

+++ scripts/kf_parallel_demo.py
@@ -84,7 +84,7 @@ if __name__ == "__main__":
     C = jnp.array([
         [1, 0, 0, 0],
         [0, 1, 0, 0]
-    ])
+    ]).astype(float)
 
     state_size, _ = A.shape
     observation_size, _ = C.shape
@@ -92,7 +92,7 @@ if __name__ == "__main__":
     Q = jnp.eye(state_size) * 0.01
     R = jnp.eye(observation_size) * 1.2
     # Prior parameter distribution
-    mu0 = jnp.array([8, 10, 1, 0])
+    mu0 = jnp.array([8, 10, 1, 0]).astype(float)
     Sigma0 = jnp.eye(state_size) * 0.1
 
     n_samples = 4
